chore: refactored login logic to hook
