+++ src/components/UserNavbarItem/item.desktop.tsx
@@ -9,15 +9,10 @@ import {
 } from '@deriv/quill-icons';
 import useLogout from '@site/src/hooks/useLogout';
 import useDeviceType from '@site/src/hooks/useDeviceType';
-import {
-  requestOidcAuthentication,
-  TOAuth2EnabledAppList,
-  useIsOAuth2Enabled,
-} from '@deriv-com/auth-client';
 
 import { IUserNavbarItemProps } from './item.types';
 import styles from './UserNavbarItem.module.scss';
-import useGrowthbookGetFeatureValue from '@site/src/hooks/useGrowthbookGetFeatureValue';
+import { useHandleLogin } from '@site/src/hooks/useHandleLogin';
 
 interface IActionProps {
   handleClick: () => void;
@@ -66,12 +61,9 @@ const DashboardActions: React.FC<IActionProps> = ({ handleClick, isDesktop }) =>
 };
 
 const SignedInActions: React.FC<IActionProps> = ({ handleClick, isDesktop }) => {
-  const [OAuth2EnabledApps, OAuth2EnabledAppsInitialised] =
-    useGrowthbookGetFeatureValue<TOAuth2EnabledAppList>({
-      featureFlag: 'hydra_be',
-    });
-
-  const isOAuth2Enabled = useIsOAuth2Enabled(OAuth2EnabledApps, OAuth2EnabledAppsInitialised);
+  const { handleLogin } = useHandleLogin({
+    onClickLogin: handleClick,
+  });
 
   const signedInButtonClasses = clsx('navbar__item', styles.UserNavbarItem, styles.SignedInButton);
 
@@ -80,14 +72,7 @@ const SignedInActions: React.FC<IActionProps> = ({ handleClick, isDesktop }) => 
       <Button
         variant='secondary'
         color='black'
-        onClick={async () => {
-          if (isOAuth2Enabled) {
-            await requestOidcAuthentication({
-              redirectCallbackUri: `${window.location.origin}/callback`,
-            });
-          }
-          handleClick();
-        }}
+        onClick={handleLogin}
         className={signedInButtonClasses}
         data-testid='sa_login'
       >
